sqLogger: don't interpolate arguments on error
If the sql query fails, don't interpolate arguments into the query in
case that covers up the reason for failing in the first place. The most
common failure mode would be using ? placeholders on Postgres, which is
invalid but after interpolation the arguments seemingly go into place
and the query passes.